bloglist user and token admin

+++ part4/bloglist/controllers/blogs.js
@@ -1,39 +1,37 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  /*Blog
-    .find({})
-    .then(blogs => {
-      response.json(blogs)
-    })*/
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const body = request.body
+  const user = request.user
 
-  const result = await blog.save()
-  response.status(201).json(result)
-  /*
-  blog
-    .save()
-    .then(result => {
-      response.status(201).json(result)
-    })
-    .catch(error => next(error)) */
-  /*try {
-    const result = await blog.save()
-    response.status(201).json(result)
-  }
-  catch (exception) {
-    next(exception)
-  }*/
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user.id
+  })
+
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === request.user.id) {
+    await Blog.findByIdAndRemove(request.params.id)
+  } else {
+    response.status(401).json({ error: 'unauthorized' })
+  }
   response.status(204).end()
 })
